requirements.txt
argparse
traceback
logging
json
requests
Bio
pyspotlight

